AnkiCardCreatorAddon/requirements.txt
gTTS
requests
deep-translator
eng-to-ipa
aqt
google.generativeai
python-dotenv